+++ src/jdbc_42/org/firebirdsql/util/Base64DecoderImpl.java
@@ -1,0 +1,36 @@
+/*
+ * Firebird Open Source JavaEE Connector - JDBC Driver
+ *
+ * Distributable under LGPL license.
+ * You may obtain a copy of the License at http://www.gnu.org/copyleft/lgpl.html
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * LGPL License for more details.
+ *
+ * This file was created by members of the firebird development team.
+ * All individual contributions remain the Copyright (C) of those
+ * individuals.  Contributors to this file are either listed here or
+ * can be obtained from a source control history command.
+ *
+ * All rights reserved.
+ */
+package org.firebirdsql.util;
+
+import java.util.Base64;
+
+/**
+ * Java 8 and higher implementation of {@link Base64Decoder}.
+ *
+ * @author <a href="mailto:mrotteveel@users.sourceforge.net">Mark Rotteveel</a>
+ */
+public class Base64DecoderImpl implements Base64Decoder {
+
+    @SuppressWarnings("Since15")
+    @Override
+    public byte[] decodeBase64(String base64) {
+        return Base64.getDecoder().decode(base64);
+    }
+
+}