+++ src/main/org/firebirdsql/gds/ng/wire/version12/Version12Descriptor.java
@@ -49,7 +49,7 @@ public final class Version12Descriptor extends AbstractProtocolDescriptor implem
                 WireProtocolConstants.arch_generic,
                 WireProtocolConstants.ptype_lazy_send, // Protocol implementation expects lazy send
                 WireProtocolConstants.ptype_lazy_send,
-                2);
+                3);
     }
 
     @Override