+++ src/main/org/firebirdsql/jca/FBManagedConnectionFactory.java
@@ -425,6 +425,16 @@ public class FBManagedConnectionFactory implements ManagedConnectionFactory, Fir
         connectionProperties.setWireCrypt(wireCrypt);
     }
 
+    @Override
+    public String getDbCryptConfig() {
+        return connectionProperties.getDbCryptConfig();
+    }
+
+    @Override
+    public void setDbCryptConfig(String dbCryptConfig) {
+        connectionProperties.setDbCryptConfig(dbCryptConfig);
+    }
+
     public int hashCode() {
         if (hashCode != 0) 
             return hashCode;