+++ src/main/org/firebirdsql/jdbc/FirebirdConnectionProperties.java
@@ -420,6 +420,7 @@ public interface FirebirdConnectionProperties {
      * Get the wire encryption level value.
      *
      * @return Wire encryption level ({@code null} implies {@code DEFAULT})
+     * @since 3.0.4
      */
     String getWireCrypt();
 
@@ -431,6 +432,24 @@ public interface FirebirdConnectionProperties {
      * </p>
      *
      * @param wireCrypt Wire encryption level
+     * @since 3.0.4
      */
     void setWireCrypt(String wireCrypt);
+
+    /**
+     * Get the database encryption plugin configuration.
+     *
+     * @return Database encryption plugin configuration, meaning plugin specific
+     * @since 3.0.4
+     */
+    String getDbCryptConfig();
+
+    /**
+     * Sets the database encryption plugin configuration.
+     *
+     * @param dbCryptConfig Database encryption plugin configuration, meaning plugin specific
+     * @since 3.0.4
+     */
+    void setDbCryptConfig(String dbCryptConfig);
+
 }