+++ src/test/org/firebirdsql/gds/ng/TestFbConnectionProperties.java
@@ -171,6 +171,14 @@ public class TestFbConnectionProperties {
     }
 
     @Test
+    public void testDbCryptConfig() {
+        assertNull(info.getDbCryptConfig());
+        final String dbCryptConfig = "ABCDEF";
+        info.setDbCryptConfig(dbCryptConfig);
+        assertEquals(dbCryptConfig, info.getDbCryptConfig());
+    }
+
+    @Test
     public void testCopyConstructor() throws Exception {
         info.setDatabaseName("testValue");
         info.setServerName("xyz");
@@ -178,6 +186,7 @@ public class TestFbConnectionProperties {
         info.setConnectionDialect((short) 2);
         info.setConnectTimeout(15);
         info.setWireCrypt(WireCrypt.REQUIRED);
+        info.setDbCryptConfig("XYZcrypt");
 
         FbConnectionProperties copy = new FbConnectionProperties(info);
         BeanInfo beanInfo = Introspector.getBeanInfo(FbConnectionProperties.class);